Modal Title, Body and Footer component created; Modal component returned by useModal hook
