+++ src/components/Modal/Modal.Footer.tsx
@@ -1,0 +1,45 @@
+/**
+ * Copyright 2024 Mia srl
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { ReactElement } from 'react'
+
+import { FooterProps } from './Modal.props'
+import styles from './Modal.module.css'
+
+const { footerButtons } = styles
+
+/**
+ * Footer component of Modal, which has within it and manages the modal footer.
+ *
+ * @returns {Footer} Modal Footer component
+ */
+export const Footer = ({
+  footer,
+}: FooterProps): ReactElement => {
+  const { buttons, extra } = footer || {}
+  return (
+    <div className={styles.footer}>
+      {footer && <>
+        <div className={footerButtons}>
+          {buttons}
+        </div>
+        {extra}
+      </>}
+    </div>
+  )
+}
